feat: :sparkles: Add string/arr lib functions
