+++ lib/Makefile
@@ -18,7 +18,7 @@ TEST     := test
 # ===== Packages =====
 PKGS     := string dict system ft_math list
 
-stringV  := new util util2 array split join
+stringV  := new util util2 array array2 split join
 dictV    := new del util hash set get expand print status
 systemV  := calloc
 ft_mathV := math
